+++ cucumber-core/src/main/java/io/cucumber/core/runner/Options.java
@@ -1,6 +1,7 @@
 package io.cucumber.core.runner;
 
 import io.cucumber.core.backend.ObjectFactory;
+import io.cucumber.core.eventbus.UuidGenerator;
 import io.cucumber.core.snippets.SnippetType;
 
 import java.net.URI;
@@ -15,5 +16,7 @@ public interface Options {
     SnippetType getSnippetType();
 
     Class<? extends ObjectFactory> getObjectFactoryClass();
+
+    Class<? extends UuidGenerator> getUuidGeneratorClass();
 
 }